# apps/api/docs/openapi.yml
openapi: "3.1.0"
info:
  title: cv-api
  version: "1.0.0"
  description: >
    HTTP/WebSocket API server for the CV management system.
    Wraps Make targets with secure REST endpoints, authentication,
    rate limiting, and streaming output.
servers:
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/actions/{target}:
    post:
      tags: [actions]
      summary: Execute Make target
      operationId: executeAction
      security:
        - cookieAuth: []
        - apiKeyAuth: []
      parameters:
        - name: target
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionRequest"
      responses:
        "200":
          description: Target completed synchronously
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "202":
          description: Target accepted and running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/targets:
    get:
      tags: [actions]
      summary: List Make targets
      description: Returns all allowed Make targets from the configuration.
      operationId: listTargets
      security:
        - cookieAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: List of allowed Make targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Target"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/api-keys:
    get:
      tags: [api-keys]
      summary: List API keys
      description: >
        Returns prefix and role information for all stored API keys.
        Full key values are never returned. Requires admin role.
      operationId: listAPIKeys
      security:
        - cookieAuth: []
        - apiKeyAuth: []
      responses:
        "200":
          description: List of API keys (prefixes only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [api-keys]
      summary: Generate API key
      description: >
        Creates a new API key and returns the full value exactly once.
        The key is added to the runtime store immediately.
        Requires admin role.
      operationId: generateAPIKey
      security:
        - cookieAuth: []
        - apiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateAPIKeyRequest"
      responses:
        "201":
          description: API key generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateAPIKeyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /ws/actions/{target}:
    get:
      tags: [streaming]
      summary: WebSocket stream
      description: >
        WebSocket connection for real-time Make target output.
        Authenticates via token query parameter or session cookie during
        handshake. Requires at least editor role.

        Protocol:
        - Client sends ActionRequest JSON as first message
        - Server sends WSMessage with type "stdout", "stderr", "exit", or "error"
        - Connection closes after the Make process exits
      operationId: wsStream
      parameters:
        - name: target
          in: path
          required: true
          schema:
            type: string
          description: Make target name to execute
        - name: token
          in: query
          schema:
            type: string
          description: JWT session token for authentication
      responses:
        "101":
          description: WebSocket upgrade successful
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cv_session
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
  schemas:
...
    ActionRequest:
      type: object
      properties:
        application:
          type: string
        args:
          type: object
          additionalProperties:
            type: string
    ActionResult:
      type: object
      required: [job_id, target, status, exit_code, duration_ms]
      properties:
        job_id:
          type: string
        target:
          type: string
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        exit_code:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
        duration_ms:
          type: integer
...
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
...
    Target:
      type: object
      required: [name, category, description]
      properties:
        name:
          type: string
          description: Make target name
        category:
          type: string
          description: Target category for grouping
        description:
          type: string
          description: Human-readable description of what the target does
        args:
          type: array
          items:
            type: string
          description: Required argument names
        timeout:
          type: string
          description: Timeout override (e.g. "5m", "30s")
...
    APIKeyInfo:
      type: object
      required: [prefix, role, added_at]
      properties:
        prefix:
          type: string
          description: First 8 characters of the API key
        role:
          type: string
          enum: [viewer, editor, admin]
          description: RBAC role assigned to this key
        added_at:
          type: string
          format: date-time
          description: When the key was created
    APIKeyListResponse:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            $ref: "#/components/schemas/APIKeyInfo"
    GenerateAPIKeyRequest:
      type: object
      properties:
        role:
          type: string
          enum: [viewer, editor]
          default: editor
          description: Role to assign to the new key (defaults to editor)
    GenerateAPIKeyResponse:
      type: object
      required: [key, prefix, role]
      properties:
        key:
          type: string
          description: Full API key value (returned exactly once)
        prefix:
          type: string
          description: First 8 characters of the key
        role:
          type: string
          enum: [viewer, editor]
          description: Role assigned to this key
...